% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2tree.R
\docType{class}
\name{json2tree}
\alias{json2tree}
\title{json2tree}
\usage{
json2tree(
  eqdt.json = NULL,
  file.output = "EQtree.txt",
  output.directory = base::paste0("json2tree.", base::gsub(x = base::Sys.time(),
    pattern = "( |:)", replacement = "-"))
)
}
\arguments{
\item{eqdt.json}{character vector of length 1, containing path to EQDT JSON file}

\item{file.output}{character vector of length 1, containing path to output file}

\item{output.directory}{character vector of length 1, path to output directory. Default : base::paste0("json2tree.",base::gsub(x=base::Sys.time(),pattern="( |:)",replacement="-")). If NULL, output.directory will be set to working directory.}
}
\description{
This function accepts a EQDT JSON file of an electronic questionnaire, and
visualizes the logic structure of the electronic questionnaire in a tree
structure, and prints that tree structure to file.
}
\examples{
# See the vignette 'json2tree-usage' for more details, by executing the command: vignette("stcEQDT-usage")
library(stcEQDT)

json2tree(
    eqdt.json = system.file("extdata", "OID-exit.json", package = "stcEQDT"),
    file.output = 'EQtree-OID-exit.txt'
    )

}
